+++ src/components/data/WDSContent.js
@@ -1,6 +1,6 @@
 import { Fragment, useState } from 'react';
 import { h } from 'react-hyperscript-helpers';
-import DataTable from 'src/components/data/DataTable';
+import DataTable from 'src/data/data-table/shared/DataTable';
 import { wdsToEntityServiceMetadata } from 'src/libs/ajax/data-table-providers/WdsDataTableProvider';
 import colors from 'src/libs/colors';
