+++ src/pages/workspaces/workspace/workflows/DataStepContent.js
@@ -4,10 +4,10 @@ import PropTypes from 'prop-types';
 import { useState } from 'react';
 import { div, h } from 'react-hyperscript-helpers';
 import { ButtonPrimary, ButtonSecondary, IdContainer, RadioButton } from 'src/components/common';
-import DataTable from 'src/components/data/DataTable';
 import { icon } from 'src/components/icons';
 import { ValidatedInput } from 'src/components/input';
 import Modal from 'src/components/Modal';
+import DataTable from 'src/data/data-table/shared/DataTable';
 import { EntityServiceDataTableProvider } from 'src/libs/ajax/data-table-providers/EntityServiceDataTableProvider';
 import { FormLabel } from 'src/libs/forms';
 import * as Style from 'src/libs/style';
